// src/types/UserAddressDataModel.ts
type UserAddressDataModel = {
    _id: string;
    fname: string;
    lname: string;
    email: string;
    streetAddress: string;
    city: string;
    state: string;
    country: string;
    postalCode: string;
    userId: string;
    active: boolean;
}


export default UserAddressDataModel